register kind of new *NodeSet objects
