+++ api/v1alpha1/nodeset_types.go
@@ -2,7 +2,6 @@ package v1alpha1
 
 import (
 	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 // NodeSetSpec describes an group nodes object
@@ -68,86 +67,4 @@ type NodeSetSpecInline struct {
 	// (Optional) Common NodeSetSpec configuration
 	// +optional
 	NodeSetSpec `json:",inline"`
-}
-
-// StorageNodeSetSpec describes an group nodes of Storage object
-type StorageNodeSetSpec struct {
-	// Name of parent Storage object to reference
-	// +required
-	StorageRef string `json:"storageRef,omitempty"`
-
-	NodeSetSpec `json:",inline"`
-}
-
-// StorageNodeSetStatus defines the observed state
-type StorageNodeSetStatus struct {
-	State      string             `json:"state"`
-	Conditions []metav1.Condition `json:"conditions,omitempty"`
-}
-
-//+kubebuilder:object:root=true
-//+kubebuilder:subresource:status
-//+kubebuilder:printcolumn:name="Status",type="string",JSONPath=".status.state",description="The status of this StorageNodeSet"
-//+kubebuilder:printcolumn:name="Age",type="date",JSONPath=".metadata.creationTimestamp"
-
-// StorageNodeSet declares StatefulSet parameters
-type StorageNodeSet struct {
-	metav1.TypeMeta `json:",inline"`
-	// +optional
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-	// +optional
-	Spec StorageNodeSetSpec `json:"spec,omitempty"`
-	// +optional
-	// +kubebuilder:default:={state: "Pending"}
-	Status StorageNodeSetStatus `json:"status,omitempty"`
-}
-
-//+kubebuilder:object:root=true
-
-// StorageNodeSetList contains a list of StorageNodeSet
-type StorageNodeSetList struct {
-	metav1.TypeMeta `json:",inline"`
-	metav1.ListMeta `json:"metadata,omitempty"`
-	Items           []StorageNodeSet `json:"items"`
-}
-
-// DatabaseNodeSetSpec describes an group nodes of Database object
-type DatabaseNodeSetSpec struct {
-	// Name of parent Database object to reference
-	// +required
-	DatabaseRef string `json:"databaseRef,omitempty"`
-
-	NodeSetSpec `json:",inline"`
-}
-
-//+kubebuilder:object:root=true
-//+kubebuilder:subresource:status
-//+kubebuilder:printcolumn:name="Status",type="string",JSONPath=".status.state",description="The status of this DatabaseNodeSet"
-//+kubebuilder:printcolumn:name="Age",type="date",JSONPath=".metadata.creationTimestamp"
-
-// DatabaseNodeSet declares StatefulSet parameters for storageRef
-type DatabaseNodeSet struct {
-	metav1.TypeMeta `json:",inline"`
-	// +optional
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-	// +optional
-	Spec DatabaseNodeSetSpec `json:"spec,omitempty"`
-	// +optional
-	// +kubebuilder:default:={state: "Pending"}
-	Status DatabaseNodeSetStatus `json:"status,omitempty"`
-}
-
-// DatabaseNodeSetStatus defines the observed state
-type DatabaseNodeSetStatus struct {
-	State      string             `json:"state"`
-	Conditions []metav1.Condition `json:"conditions,omitempty"`
-}
-
-//+kubebuilder:object:root=true
-
-// DatabaseNodeSetList contains a list of DatabaseNodeSet
-type DatabaseNodeSetList struct {
-	metav1.TypeMeta `json:",inline"`
-	metav1.ListMeta `json:"metadata,omitempty"`
-	Items           []DatabaseNodeSet `json:"items"`
 }
